+++ xnum/params.py
@@ -72,8 +72,9 @@ TAI_THAM_THAM_DIGITS = ['᪐', '᪑', '᪒', '᪓', '᪔', '᪕', '᪖', '᪗', 
 NYIAKENG_PUACHUE_HMONG_DIGITS = ['𞅀', '𞅁', '𞅂', '𞅃', '𞅄', '𞅅', '𞅆', '𞅇', '𞅈', '𞅉']
 MALAYALAM_DIGITS = ['൦', '൧', '൨', '൩', '൪', '൫', '൬', '൭', '൮', '൯']
 BRAHMI_DIGITS = ['𑁦', '𑁧', '𑁨', '𑁩', '𑁪', '𑁫', '𑁬', '𑁭', '𑁮', '𑁯']
-MYANMAR_TAI_LAING_DIGITS = ['꧰','꧱','꧲','꧳','꧴','꧵','꧶','꧷','꧸','꧹']
+MYANMAR_TAI_LAING_DIGITS = ['꧰', '꧱', '꧲', '꧳', '꧴', '꧵', '꧶', '꧷', '꧸', '꧹']
 AHOM_DIGITS = ['𑜰', '𑜱', '𑜲', '𑜳', '𑜴', '𑜵', '𑜶', '𑜷', '𑜸', '𑜹']
+WARANG_CITI_DIGITS = ['𑣠', '𑣡', '𑣢', '𑣣', '𑣤', '𑣥', '𑣦', '𑣧', '𑣨', '𑣩']
 
 NUMERAL_MAPS = {
     "english": ENGLISH_DIGITS,
@@ -145,7 +146,8 @@ NUMERAL_MAPS = {
     "malayalam": MALAYALAM_DIGITS,
     "brahmi": BRAHMI_DIGITS,
     "myanmar_tai_laing": MYANMAR_TAI_LAING_DIGITS,
-    "ahom": AHOM_DIGITS
+    "ahom": AHOM_DIGITS,
+    "warang_citi": WARANG_CITI_DIGITS,
 }
 
 ALL_DIGIT_MAPS = {}
@@ -227,6 +229,7 @@ class NumeralSystem(Enum):
     BRAHMI = "brahmi"
     MYANMAR_TAI_LAING = "myanmar_tai_laing"
     AHOM = "ahom"
+    WARANG_CITI = "warang_citi"
     AUTO = "auto"